Generate_from_sysctl_list refactoring. Minor changes and fixes in paths.
