+++ diffkemp/building/building_snapshot_files/build_utils.py
